api/rest/urls.py
from rest_framework.routers import DefaultRouter
from django.urls import path, include
from . import views
from rest_framework_simplejwt.views import (
    TokenObtainPairView,
    TokenRefreshView,
    TokenVerifyView
)

router = DefaultRouter()
router.register('post', views.PostModelViewSet)

urlpatterns = [
    path('', include(router.urls))
]